+++ CHANGELOG.rst
@@ -5,9 +5,14 @@ Changelog
 0.2.0 (unreleased)
 ==================
 
+Features:
+
 * Add framework-independent ``HyperlinkRelated`` field.
-* Fix for making ``HyperlinkRelated`` read-only by defualt.
 * Support inflection of attribute names via the ``inflect`` class Meta option.
+
+Bug fixes:
+
+* Fix for making ``HyperlinkRelated`` read-only by defualt.
 
 0.1.0 (2015-09-15)
 ==================